refactor: split index in children components"
